🚚 move definitions of noise-aware DD package configs to main library
